[Pal] db_rtld.c: Fix GDB on PIE executables by fixing mapstart/mapend
Previously, there was a latent bug in db_rtld.c: add_elf_object() --
the mapstart variable was incorrectly initialized to zero. This led to
PIE executables (which typically have two LOAD segments, the code
segment starting with 0x0 and the GOT/PLT/dynamic segment starting with
e.g. 0x200000) to incorrectly set mapstart to the higher-LOAD-segment
address (and not the code address). This in turn confused GDB which
tried to add executable symbols from the wrong offset.

This commit fixes this bug by initializing mapstart to the highest
possible value ("~0") and removing the check for zero. This is similar
to what the Glibc code does. Also, mapend is fixed in a similar way.